+++ src/lib/containers/__tests__/works-for-you-tests.tsx
@@ -65,6 +65,11 @@ let notificationsResponse = () => {
               date: "Mar 16",
               message: "1 Work Added",
               artworks: [ { title: "Anti-Product Postcard" } ],
+              image: {
+                resized: {
+                  url: "cloudfront.url",
+                },
+              },
             },
           },
           {
@@ -73,6 +78,11 @@ let notificationsResponse = () => {
               date: "Mar 16",
               message: "2 Works Added",
               artworks: [ { title: "Corazón de Roca con Sangre" }, { title: "Butterfly" } ],
+              image: {
+                resized: {
+                  url: "cloudfront.url",
+                },
+              },
             },
           },
         ],