+++ src/lib/containers/works_for_you.tsx
@@ -27,6 +27,7 @@ interface Props extends RelayProps {}
 interface State {
   dataSource: ListViewDataSource | null
   sideMargin: number
+  topMargin: number
 }
 
 export class WorksForYou extends React.Component<Props, State> {
@@ -41,8 +42,8 @@ export class WorksForYou extends React.Component<Props, State> {
     this.state = {
       dataSource,
       sideMargin: 20,
+      topMargin: 0,
     }
-
   }
 
   componentDidMount() {
@@ -65,7 +66,9 @@ export class WorksForYou extends React.Component<Props, State> {
   onLayout = (event: LayoutChangeEvent) => {
     const layout = event.nativeEvent.layout
     const sideMargin = layout.width > 600 ? 40 : 20
-    this.setState({ sideMargin })
+    const topMargin = layout.width > 600 ? 20 : 0
+
+    this.setState({ sideMargin, topMargin })
   }
 
   render() {
@@ -93,7 +96,7 @@ export class WorksForYou extends React.Component<Props, State> {
                 renderSeparator={(sectionID, rowID) =>
                   <View key={`${sectionID}-${rowID}`} style={styles.separator} /> as React.ReactElement<{}>
                 }
-                style={{marginTop: 20}}
+                style={{marginTop: this.state.topMargin}}
       />)
   }